refactor(#723): move pure matching helpers and _chunked_gather out of the orchestrator
Verbatim relocation of the matching predicates/filters/floors to lookup/matching.py and the _chunked_gather API-call-cap machinery to lookup/concurrency.py, per the decomposition manifest in #722. The orchestrator imports what its remaining code still consumes; no logic changes. Review this commit with git diff --color-moved=zebra.
